Trying to add avatars and handle an error

+++ cogs/ttt.py
@@ -31,21 +31,25 @@ class ttt(commands.Cog):
     @commands.command(  name = 'ttt_new',
                         help = '>ttt_new @challenged',
                         brief = '- Start a new game of Tic-Tac-Toe with someone.')
-    async def ttt_new_game(self, ctx, challenged: discord.User):
+    async def ttt_new_game(self, ctx, challenged: discord.User = None):
+        if challenged == None:
+            await ctx.send("You need to challenge someone.\n >ttt_new @challenged")
+
+            return
+
         await ctx.message.delete()
         
         try:
             self.player2 = ctx.author
             self.player1 = challenged
 
-            await ctx.send(f"{self.player1.display_avatar}")
-            await ctx.send(self.player1.display_avatar)
-            await ctx.send(f"{self.player2.display_avatar}")
-            await ctx.send(self.player2.display_avatar)
+            await ctx.send(str(self.player1.display_avatar))
+            await ctx.send(str(self.player2.display_avatar))
             
             await ctx.send(f"{self.player1.mention} has been challenged to a Tic-Tac-Toe game by {self.player2.mention}")
             await ctx.send(f"https://tenor.com/view/yu-gi-oh-duel-yugi-anime-gif-7357665")
-            embed = discord.Embed(title = f"Tic-Tac-Toe {self.player1.name} vs {self.player2.name}")
+            embed = discord.Embed(title = f"Tic-Tac-Toe {self.player1.display_avatar}{self.player1.name}\
+                                             vs {self.player2.name}{self.player2.display_avatar}")
             # embed.add_image(icon_url = self.player1.display_avatar)
             # embed.add_image(icon_url = self.player2.display_avatar)
             embed.add_field(name = "Game:",
@@ -58,7 +62,7 @@ class ttt(commands.Cog):
                                     ",
                             inline = False
                             )
-            embed.set_thumbnail(url = "https://e7.pngegg.com/pngimages/923/206/png-clipart-tictactoe-tic-tac-toe-oxo-tac-tic-toe-computer-icons-games-buttons-game-logo-thumbnail.png")
+            embed.set_thumbnail(url = "https://www.emulatorpc.com/wp-content/uploads/2023/02/tic-tac-toe-on-pc.jpg.webp")
 
             await ctx.send(embed = embed)
         except Exception as e:
